+++ plc_sketchbook/stage_control_PID_ONLY/h_path.ino
@@ -2,7 +2,7 @@
 // Calculate Path
 // ---------------
 
-uint32_t calculate_path()  {
+float calculate_path()  {
 
   // Data Pre-Processing //
   // Check for Danger
@@ -10,17 +10,18 @@ uint32_t calculate_path()  {
   // Calculate Error
   // Always wants to go to desired_offset for PID only, so that distance is error
   error[histPointer] = distance_hist[histPointer] - desired_offset;
+
  
   // PID & Feed-Forward Tuning
-  uint32_t Der_error = (error[last_pointer] - error[histPointer]) / (dt);
+  float Der_error = (error[last_pointer] - error[histPointer]) / (dt);
   Int_error = error[last_pointer] * dt + Int_error;
   
-  uint32_t P_Effort = P * error[histPointer];
-  uint32_t I_Effort = I * Int_error;
-  uint32_t D_Effort = D * Der_error;
+  float P_Effort = P * error[histPointer];
+  float I_Effort = I * Int_error; // BE SURE TO ADD RE-ZEROING BEFORE USE
+  float D_Effort = D * Der_error;
 
-  // Calculate Command
-  uint32_t requested_control = command_hist[histPointer] + P_Effort + I_Effort + D_Effort;
+  // Calculate Command //NOTE THIS IS NOW IN COUNTS
+  float requested_control = command_hist[histPointer] + P_Effort + I_Effort + D_Effort;
   command_hist[histPointer] = requested_control;
   return requested_control;
 }
